client/requirements.txt
alabaster==0.7.12
astroid==2.3.2
Babel==2.7.0
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
cx-Freeze==6.0
docutils==0.15.2
entrypoints==0.3
flake8==3.7.9
idna==2.8
imagesize==1.1.0
isort==4.3.21
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==19.2
Pillow==6.2.1
pkginfo==1.5.0.1
pycodestyle==2.5.0
pycryptodomex==3.9.0
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.4.3
pyparsing==2.4.2
PyQt5==5.13.0
PyQt5-sip==12.7.0
pyqt5-tools==5.13.0.1.5
python-dotenv==0.10.3
pytz==2019.3
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.10
tabulate==0.8.5
tqdm==4.37.0
twine==2.0.0
typed-ast==1.4.0
urllib3==1.25.6
webencodings==0.5.1
wrapt==1.11.2
